Genereert soorten(groep)lijst(en) LSVI op basis van TaxongroepID

Deze functie genereert soortenlijsten (met wetenschappelijke en Nederlandse namen) uit de databank met de criteria en indicatoren voor de bepaling van de Lokale Staat van Instandhouding. Het is in feite een hulpfunctie die voor verschillende andere functies gebruikt wordt en die de complexe zoekfunctie in de tabellen met soorten uitvoert op basis van een opgegeven TaxongroepId (en in die zin iets minder gebruiksvriendelijk is). Voor een selectie van soortenlijsten op basis van specifieke parameters is de functie geefSoortenlijst() een beter alternatief.

Deze functie geeft standaard voor de gespecifieerde taxongroepen per groep een lijst van alle taxa zoals ze in de LSVI-habitatfiche vermeld zijn (genusniveau, soortniveau, subsoort,...). Op basis van de parameter soortenlijsttype kan ook gekozen worden om een volledige lijst te geven van deze taxa en alle taxa die hieronder vallen (en opgenomen zijn in de onderliggende databank).

geefSoortenlijstVoorIDs(Taxongroeplijst, Taxonlijsttype = c("LSVIfiche",
  "alle"), ConnectieLSVIhabitats = NULL)

Arguments

Taxongroeplijst
string waarin de TaxongroepId's na elkaar weergegeven worden, gescheiden door een komma. Eventueel mag dit ook een vector zijn van TaxongroepId's.

Taxonlijsttype
"LSVIfiche" betekent dat de taxonlijst van de habitatfiche wordt overgenomen, "alle" betekent dat alle soorten en alle taxonomische groepen worden weergegeven die volledig in de groepen vallen die aan de parameters voldoen.

ConnectieLSVIhabitats
Connectie met de databank met indicatoren voor de LSVI van habitats, in te stellen d.m.v. functie connecteerMetLSVIdb.

Value

Deze functie geeft een tabel met velden TaxongroepId, evt. Beschrijving, WetNaam, WetNaamKort en NedNaam (waarbij Beschrijving een omschrijving is voor een groep van taxons binnen eenzelfde indicator). WetNaam is de volledige Latijnse naam inclusief auteursnaam, WetNaamKort geeft de verkorte naam zonder auteursnaam.

Examples

# Omwille van de iets langere lange duurtijd van de commando's staat bij
# onderstaande voorbeelden de vermelding 'dontrun' (om problemen te vermijden
# bij het testen van het package). Maar de voorbeelden werken en kunnen zeker
# uitgetest worden.

# NOT RUN {
maakConnectiePool()
geefSoortenlijstVoorIDs("434,88,565")
geefSoortenlijstVoorIDs("434,88,565","alle")
library(pool)
poolClose(ConnectiePool)
# }