zet ingevoerde gegevens om naar een interval

Deze functie zet ingevoerde gegevens van meerdere types om naar een interval bestaande uit minimumwaarde en maximumwaarde. De functie gebruikt Type, Eenheid en Invoertype om te bepalen welke omzetting eventueel nodig is. Percentages worden bv. omgezet naar een decimaal getal (waarbij minimum en maximum dezelfde waarde zullen krijgen), en categorische variabelen met Invoertype Tansley worden omgezet naar de onder- en bovengrens die in de databank gegeven worden voor de betreffende categorie (op basis van parameter LIJST). Om een onderscheid te maken tussen numerieke waarden en aan-/afwezigheid, wordt voor deze laatste enkel de minimumwaarde ingevoerd (dus maximum = NA).

vertaalInvoerInterval(Dataset, LIJST, ConnectieLSVIhabitats)

Arguments

Dataset
dataframe met velden Rijnr, Type, Waarde, Eenheid en Invoertype

LIJST
Dataframe met lijst die weergeeft hoe de vertaling moet gebeuren van categorische variabelen naar numerieke waarden (en omgekeerd). Default worden deze waarden uit de databank met LSVI-indicatoren gehaald d.m.v. de functie vertaalInvoerInterval(). Aangeraden wordt om deze default te gebruiken (dus parameter niet expliciet invullen), of deze waar nodig aan te vullen met eigen schalen. Omdat er ook een omzetting moet gebeuren voor grenswaarden uit de databank, kan het niet doorgeven van een gedeelte van deze lijst problemen geven.

ConnectieLSVIhabitats
Connectie met de databank met indicatoren voor de LSVI van habitats, in te stellen d.m.v. functie connecteerMetLSVIdb.

Value

Dataframe met velden Rijnr, Min en Max